% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-rmd-yaml.R
\name{change_rmd_yaml}
\alias{change_rmd_yaml}
\title{Funcao para alterar o YAML de um rmd}
\usage{
change_rmd_yaml(input_file, ..., output_file)
}
\arguments{
\item{input_file}{arquivo de entrada}

\item{...}{o que queremos alterar.}

\item{output_file}{arquivo de saida. opcional. se não informar,
ele mostra o resultado no console}
}
\value{
o conteúdo do arquivo alterado
}
\description{
Funcao por r2evans, postado no stackoverflow (thanks!)
https://stackoverflow.com/questions/62095329/how-to-edit-an-r-markdown-yaml-header-programmatically
}
