% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_github_files.R
\name{get_default_branch}
\alias{get_default_branch}
\title{Get default branch of a GitHub repository}
\usage{
get_default_branch(repo, owner = "curso-r")
}
\arguments{
\item{repo}{Repository name (excluding owner).}

\item{owner}{GitHub organization (default is \code{curso-r}).}
}
\value{
A string.
}
\description{
Get default branch of a GitHub repository
}
\examples{
get_default_branch("main-web-scraping")

}
