% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_github_files.R
\name{list_github_files}
\alias{list_github_files}
\title{List files in a GitHub repository}
\usage{
list_github_files(repo, path = NULL, regexp = NULL, owner = "curso-r")
}
\arguments{
\item{repo}{Repository name (excluding owner).}

\item{path}{Subdirectory inside repo.}

\item{regexp}{A regular expression passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{owner}{GitHub organization (default is \code{curso-r}).}
}
\value{
A string vector of paths.
}
\description{
List files in a GitHub repository
}
\examples{
list_github_files("main-web-scraping", "slides/", "[0-9]-.*html$")

}
