% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-rmd-yaml.R
\name{change_rmd_yaml}
\alias{change_rmd_yaml}
\title{Update YAML front matter of a RMarkdown file}
\usage{
change_rmd_yaml(path, ..., new_path = NULL)
}
\arguments{
\item{path}{Path to file.}

\item{...}{Name-value pairs to update in front matter.}

\item{new_path}{Path to output file. If \code{NULL}, prints output to console.}
}
\value{
A string vector with the updated content.
}
\description{
Update YAML front matter of a RMarkdown file
}
