% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_github_files.R
\name{list_github_files}
\alias{list_github_files}
\title{List files in a GitHub repository}
\usage{
list_github_files(
  repo,
  dir = NULL,
  pattern = NULL,
  org = "curso-r",
  pat = FALSE
)
}
\arguments{
\item{repo}{Repository name.}

\item{dir}{Folder name.}

\item{pattern}{Regex pattern to look for.}

\item{org}{GitHub organization (default is \code{curso-r}).}

\item{pat}{Add PAT as request header. This reduces the chances of getting an
empty response.}
}
\value{
A string vector.
}
\description{
List files in a GitHub repository
}
\examples{
list_github_files("main-web-scraping", "slides/", "[0-9]-.*html$")

}
