% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_github_files.R
\name{list_github_files}
\alias{list_github_files}
\title{listar arquivos em um repositorio no github}
\usage{
list_github_files(
  repo,
  dir = NULL,
  pattern = NULL,
  org = "curso-r",
  pat = FALSE
)
}
\arguments{
\item{repo}{nome do repositorio}

\item{dir}{nome do diretorio}

\item{pattern}{padrao para buscar arquivos}

\item{org}{organizacao no GitHub. Por padrao eh a curso-r.}

\item{pat}{adicionar github pat como header da requisicao. Isso reduz
as chances da requisicao vir vazia. Por padrao \code{FALSE}.}
}
\value{
um vetor
}
\description{
listar arquivos em um repositorio no github
}
\examples{
list_github_files(repo = "main-web-scraping", dir = "slides/", pattern = "[0-9]-.*html$")
}
